Implemented memory reduction techniques.
Mostly this involved putting static strings into program space and
conditionally compiling debug output. Some of the code that used
to be libraries is now compiled with the main sketch. There were
a couple of reasons for this: with the inclusion of the pspc_support
library, the compile command line got too large and I was unable
to compile the sketch. Reducing the number of libraries fixed that
problem. The libraries that I moved into the main source of the
sketch were actually more appropriate with the sketch than as a
library anyway (subjective judgement). The current behavior is
now more consistent.